feat: 만료 섹션을 위한 캐시, 칭호, 유니온 아티팩트 및 펫 데이터 처리 로직 추가
